Remove generated allComponents and nonRegisteredComponents files
We had a pair of generated files, allComponents and
nonRegisteredComponents, which import all files that contain JSS styles
(ie that call registerComponent or defineStyles). Previously, we used
this to make a static stylesheet. After the static stylesheet was
removed, the last remaining usage of these files was for a unit-test
which checks all JSS styles for signs of being black-on-black in dark
mode.

Modifies the unit test to not rely on these generated files (it walks
the filesystem and does dynamic require instead, which seems to work
fine inside of jest), then removes all remnants of their existence.
